requirements.txt
Django==2.2.17
python-dotenv==0.15.0
psycopg2-binary==2.8.6
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5